requirements.txt
python-dateutil
python-json-logger
requests
fastapi[standard]
pydantic-settings
python-slugify
uvicorn
mistune
tenacity
